Added a variant of 'call_repl1' to demonstrate the usage of trace events.
Basically, it is almost the same example of call replacement in runtime
(see 'call_repl1'). This variant, however, makes use of trace events to
output data from the replacement functions to the user space rather than
just dumps everything to the system log.

See README file for the instructions on how to use the example.