"kedr_gen" is now built with CMake
It can be configured where "kedr_gen" will be installed.
In addition, if it is not a temporary installation (i.e. not for building
KEDR only), a symlink to "kedr_gen" is created in
${KEDR_INSTALL_PREFIX_EXEC} directory.

Some changes have also been made to "kedr_gen" itself to make it build
properly.

Building and installation of the parts of MiST Engine distribution that are
not necessary for KEDR are now turned off.