Fixed: mutex_lock & Ko are not exported by some kernels
Debug variants of the kernels (namely, those with CONFIG_DEBUG_LOCK_ALLOC=y)
have *_nested functions instead of mutex_lock(), etc. This was not handled
by KEDR build system and was not taken into account by "Counters" example.

For now, I have just marked these mutex functions as optional and adjusted
the example and the tests.

We probably need to improve (or maybe completely redesign) the system that
detects the presence of the functions to avoid such problems in the future.