leak_check: integrated LeakCheck into KEDR build and install system
Among other things, the source code of the payload module (namely, the file
containing the definitions of the replacement functions) is now created from
pieces, each of which corresponds to a particular target function.

The system is similar to how payload modules for call monitoring and fault
simulation are created.