doc: Disabled generation of "title" attributes and, hence, the tooltips
By default, when generating HTML from DocBook XML with the standard XSL
stylesheets, "title" attribute was generated for the div areas corresponding
to the sections, etc. Because of this, the browsers showed the tooltips when
the mouse pointer was hovering over these areas, which was really annoying.

To fix this, I have disabled generation of "title" attributes.