Added implicit handling of kfree_rcu
The availablity of kfree_rcu is now checked during the configuration phase.
If kfree_rcu is available, call_rcu() and/or call_rcu_sched are intercepted.

LeakCheck analyzes the calls to call_rcu*() and can treat some of these
calls as kfree() (see the comment in sources/leak_check/call_rcu.data and
the description of kfree_rcu() in the kernel sources for details).

In addition, call monitoring is now available for call_rcu*().