Revisited the way the information about the target functions is handled
Before this commit, each tool provided by KEDR framework stored its own copy
of that information. This duplication made maintenance more complex, same
for the addition of the new target functions.

If the information about a function became known only on the configuration
stage, things got even worse: it was needed to repeat the respective
configuration commands for each KEDR tool that used the function.

Now the information about the functions that can be processed by the tools
provided with KEDR is stored in a single place, namely, in
"sources/functions/" tree. Logically, it can be considered as a "function
database" containing function names, return types, parameter names and
types.

The standard payload modules specify which target functions they are able
to process. Only the functions processed by one or more payload module will
be configured, checked for presence on a given system.

The tests checking that KEDR can intercept the calls to the particular
functions are now also added only for the functions that are actually used
by the standard payload modules.

Another important change is that these tests are no longer tied to call
monitoring subsystem and will be available even if that substsrem has been
disabled.

In general, if KEDR is built, say, with only LeakCheck and fault simulation
support enabled, interception of schedule() function, for instance, will not
be checked by the tests because this function is not used by these tools.

Many components of the build system have beed refactored during the process
as well.