fsim: Indicators are now loaded by default when KEDR is loaded
The configuration file for fault simulation provided by KEDR now includes
the instructions not only to load the appropriate subsystem module and the
plugins but also all the standard indicators.

It is too easy for a user to forget to load an indicator module when using
fault simulation, so this change makes the common use case simpler.