leak_check: make sure flushing has completed when it is expected to
Fixes issue #28.

When a user requested LeakCheck to flush the data (echo 1 > .../flush),
the appropriate operation was queued to be performed later. As a result,
flush operation was unreliable. If the user tried to read the results
soon enough, it was possible that flushing had not completed yet and
incomplete results were shown.

On the other hand, flush_workqueue(lc->wq) was rather useless before
the calls to klc_do_flush(lc). I have moved it after such calls to make sure
all operations queued so far (including the flush operation itself)
have completed before writing to "flush" file returns.