Revisited the build system and the directory layout
1. The directory layout for the kernel-mode components of KEDR
(see "kernel" subdirectory) was changed to make it closer to
that of Livepatch in the mainline kernel 4.7 and newer:

   * kernel/kedr/ - the source of the core module, other modules may go
     there too in the future;
   * include/linux/kedr.h - the main header file, arch-independent
     definitions. #includes the header with the arch-specific stuff, see
     below;
   * arch/x86/include/asm/kedr.h - arch-specific definitions.

This makes things easier for the kernel developers to understand.

kernel/Makefile.mk (generated from Makefile.mk.in by CMake) handles the
build of the kernel modules. Among other things, it makes CONFIG_KEDR_*
variables available to the Makefiles of these modules, like the kernel
configuration parameters are available there.

Running "make" in the build tree will build the modules for the current
kernel.

If you need to rebuild the modules for another kernel, you can do the
following in the build tree:

cd kernel
make -f Makefile.mk clean
make -f Makefile.mk KBUILD_DIR=/lib/modules/<target_kernel_version>/build

With the current build system and the directory layout, it is now also
easier to prepare DKMS-based and similar packages for these kernel modules.
See the notes for the package maintainers in the main CMakeLists.txt file
for more info on that matter.

2. It is now possible to specify ARCH and CROSS_COMPILE options when
running CMake (-DARCH=<...> -DCROSS_COMPILE=<...>). If set, they will be
passed to "make" when building the kernel modules and will have the
usual meaning there.

3. It is now possible to disable building of the instrumentation
facilities (GCC plugin). Pass -DKEDR_I13N=off to cmake for that purpose.
May help when debugging the build system, etc.

4. Cosmetic fixes.