i13n: Revisited the functions used for the instrumentation of the calls
The GCC plugin now inserts the calls to thunk functions (see
kedr_stubs.c), rather than the calls to the handler stubs.

The thunks abstract the signatures of the target functions from the KEDR
core: which parameters are passes, in what order, how to get the size of
the allocation from these parameters, etc. The thunks retrieve the
needed data and just report the events to the core: "alloc" and "free"
at the moment.

Unfortunately, I did not find a way to inline these thunks or make them
built-in like TSan does, see
http://stackoverflow.com/questions/39279024/how-to-add-a-builtin-function-in-a-gcc-plugin

Still, they should be OK as they are for now.

See the comments in kedr_stubs.c for the guarantees the handlers may rely
upon.