i13n: allocate and free kedr_local instances in the thunks
...rather than in the handlers. If the thunks manage kedr_local objects,
this is easier to implement: the thunks are always called while the
handlers can come and go at any moment.

Besides, let us simply return NULL from fentry thunk if it has failed to
allocate the kedr_local.

Let us also require that the instrumented code does not dereference the
pointer to kedr_local. Only the thunks and the handlers may do so.

This way, we no longer need 'valid' flag in kedr_local.