core: implemented setting and removing function handlers via Ftrace
The ideas and part of the code are borrowed from the implementation of
livepatch in the mainline kernel.

The instrumented parts of the kernel contain the calls to the special stubs.
The stubs do nothing by themselves but can be redirected to the real
handlers with the help of Ftrace.

When the core module (kedr.ko) is loaded, it finds the stubs by looking
through kallsyms and uses Ftrace to "replace" them with real handlers
that it provides. When the core module is unloaded, the handlers are
detached and the instrumented code will continue working, calling the
stubs that do nothing, etc.