core: separate enabling/disabling KEDR from loading/unloading kedr.ko
Besides:

1. Do not ignore the errors kedr_disable() may report. Ftrace may fail to
detach its handler, etc., consequently, kedr_disable() may fail.

2. While KEDR is enabled, do not allow kedr.ko to be unloaded.

Before this commit, the core module could be unloaded in that case
anyway and the kernel would crash when the target module tried to call
the handlers.

kedr.ko now works as follows.

When it is loaded, KEDR is disabled by default. The user should enable
it explicitly when needed:

  echo 1 > /sys/kernel/kedr/enabled

... and disable when done:

  echo 0 > /sys/kernel/kedr/enabled

kedr.ko can be unloaded only when KEDR is disabled.

I am still not quite happy with the error handling in
kedr_enable/kedr_disable and the related code (one cannot reliably
detach Ftrace handlers) but it should be OK for now.