core: Allow resolution of code addresses in a module after it is unloaded
KEDR tools may store the addresses in the kernel code where some events
has happened and process the addresses later. If the code is from a
module and the module is unloaded before KEDR outputs the info about the
event, we cannot simply use printk with "%pS" to resolve the address
(get the symbol it belongs to and the offset).

This is also the case for the events from the init code of a module
which is unloaded from memory after module_init() finishes.

In KEDR 0.x, LeakCheck resolves all stored addresses that belong to a
module when that module is unloaded. This might be inefficient: KEDR
actually needs to resolve the addresses in the backtraces for the
memory leaks reported to the user rather than all addresses.

We use a different approach here: store the "layered" map of the code
areas of the kernel modules. "Layered" means KEDR keeps the history of
loaded modules there too. When a module is loaded, a new layer is
created in the module map that contains the information of the init and
core areas for this module. The previos layers remain intact.

When storing an event, KEDR can store the value of 'current_modmap', the
map of the code areas at the moment, somewhere as well. This does not
need additional locking.

Later, if it is needed to resolve the code address in the event,
kedr_resolve_address() (defined in modmap.c) can be called to do thi
properly. Note that it needs 'kedr_mutex' locked.

kedr_resolve_address() does not determine the function the address
belongs to, only the code area, e.g.,

  [<ffffffffba03f025>] core+0x25 [kedr_common_target]
  [<ffffffffba0350de>] init+0xde [kedr_common_target]

for the addresses in the kernel module, and

  [<ffffffffa31e7370>] core+0x1e7370

for an address in the kernel.

A userspace tool (TBD) may then use the binaries of the modules, the
debug info and whatever else to resolve the output code locations
further. This can be more convenient than using %pS, by the way, in case
there are several functions with the same name in the kernel or a
module. The form "init|core+0xoffset" is non-ambiguous.

Unresolved addresses are output as is:

  [<ffffffffe31e7370>] 0xffffffffe31e7370