Revisited the instrumentation facilities
Using GIMPLE to create the code to be inserted before and after the target
function calls is difficult, error-prone and hard to maintain.

Instead, I have moved the instrumentation rules into a human-readable
file, i13n/rules.yml. The format is described in the comments there. The
thing is that you write the rules with sequences of C-like statements
and the GCC plugin will translate them into GIMPLE for you.

The rules can not only add the calls to stubs before/after the target
function calls. They can add the calls to helper functions to prepare
the arguments for the handlers. This way, the stubs can really do
nothing and the handlers from KEDR core get "pre-processed" events like
"alloc", "free", etc.

It is also no longer needed to store some of the data in struct
kedr_local when it can be prepared by the helpers, etc.

The definitions and declarations of the stubs are now generated
from kernel/kernel/kedr/core.c automatically during the build. That is,
kedr_helpers.{c,h} are now generated.