Require Meson 0.46+ and query the right compiler for plugin headers
Starting from v0.46, Meson allows passing a compiler object to
run_command(). This, in turn, allows the build script to query the
currently selected compiler rather than the default GCC for the path to
the plugin development files.

This is helpful if the system has 2 or more versions of GCC installed and
it is needed to use a non-default one.