tests: made it possible to check the events reported to dmesg
And added such checks to the "basics" test.

When kedr.ko is loaded with debug=1, it will report to dmesg the events
("alloc", "free") it has captured. Let us make the tests able to check
if KEDR captured all the expected events.

The expected events are specified in a file, example:
tests/events_basics.txt. See the comments in that file for the format.

The functionality is implemented in KedrTest class. If a subclass calls
check_dmesg_after() it may pass the path to the file with the expected
events there as 'events_file' argument. If it is not specified,
check_dmesg_after() will only check dmesg for kernel BUGs and WARNINGs,
as before.