tests: added a test for the instrumentation with the custom rules.
The test checks if the instrumentation with particular constructs in
the rules is performed correctly. See rules_custom.yml for details on
which kinds of rules are tested here.

This test is mostly for the rules which are unlikely to be used, e.g.
the rules where a variable gets assigned to twice, where a constant is
passed to a handler or a helper. etc.

To be able to check if the correct events have been reported to dmesg,
I added support for explicit address values in tests/events_*.txt:
the events like "alloc 4000 0x10", etc.