tests: added a test to check detection of alloc/free events
See KedrTestEvents test case in tests/test_kedr.py. The test loads a
special kernel module and then makes is call different functions that
allocate and/or free memory. It then checks the events reported to dmesg
(debug mode is on so they should be reproted there).

Note that some of the subtests of this test may fail, because handling
of the corresponding functions is not yet implemented, to be done later:

        krealloc.01
        krealloc.02
        krealloc.03
        krealloc.04
        __krealloc.03
        kmemdup.01
        memdup_user.01
        strndup_user.01
        memdup_user_nul.01
        kasprintf.01
        kvasprintf.01
        kfree_rcu.01

...and may be a few more.