i13n: take function renaming because of CONFIG_FORTIFY_SOURCE into account
If CONFIG_FORTIFY_SOURCE is set in the kernel, additional checks are made
in the string functions. For example, kmemdup() becomes inline, checks for
potential overflows and the calls __real_kmemdup() - it is this call that
the plugin may see here. __real_kmemdup() is then renamed into kmemdup()
again, but it seems to be done at a later stage.

See the mainline commit 6974f0c4555e "include/linux/string.h:
add the option of fortified string.h functions".

As a workaround, let us check if the function has "__real_" prefix and if
so, lookup the rule set for the name without that prefix.